requirements.txt
Flask==2.3.2
groq==0.5.1
python-dotenv==1.0.0
